\name{insilico.digest.internal}
\alias{insilico.digest.internal}

\title{
Internal function to digest DNA sequence using one restriction enzyme. 
}
\description{
Perform an in silico digestion of a DNA sequence using one restriction enzyme. For internal use, not to be called by the user.
}
\usage{
insilico.digest.internal(DNAseq, recognition_code, cut_site_5prime, cut_site_3prime)
}

\author{
Jason Weir
}
\seealso{\code{\link{sim.DNAseq}}, \code{\link{ref.DNAseq}}.}


\keyword{ internal }
